% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_ae.R
\name{ae1}
\alias{ae1}
\title{Create Adverse event list.}
\usage{
ae1(aet)
}
\arguments{
\item{aet}{the dataset created by \code{create_aet()}}
}
\value{
a data frame
}
\description{
ae1
}
\seealso{
\code{\link{create_aet}}
}

