% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_ae.R
\name{ae2}
\alias{ae2}
\title{list Adverse Envent 2.}
\usage{
ae2(aet)
}
\arguments{
\item{aet}{the dataset created by \code{create_aet()}}
}
\value{
a data frame
}
\description{
ae2
}
\seealso{
\code{\link{create_aet}}
}

