% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_ae.R
\name{ae3}
\alias{ae3}
\title{list Adverse Envent Non Drug Therapy.}
\usage{
ae3(aet)
}
\arguments{
\item{aet}{created by \code{create_aet()}}
}
\value{
a data frame
}
\description{
ae3
}
\seealso{
\code{\link{create_aet}}
}

