% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_ae.R
\name{ae_listing}
\alias{ae_listing}
\title{list Adverse Envent}
\usage{
ae_listing(aet, type = 1)
}
\arguments{
\item{aet}{created by \code{create_aet}}

\item{type}{an indicator. Should I list \code{ae1}, \code{ae2} or \code{ae3}?}
}
\value{
a data frame
}
\description{
list the ae
}
\seealso{
\code{\link{create_aet}}
}

