% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_ae.R
\name{assign_ae_trt}
\alias{assign_ae_trt}
\title{Assing AE to treatment}
\usage{
assign_ae_trt(ae, ex)
}
\arguments{
\item{ae}{the dataset ae read from sas}

\item{ex}{the dataset ex read from sas}
}
\value{
a data frame with one more column specifying which treatment the AE should be in.
}
\description{
Assign ae to a treatment.
}

