% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_quality.R
\name{check_data}
\alias{check_data}
\title{check the replicates and issue a message if there's issue}
\usage{
check_data(data, reps = NULL, printed = TRUE)
}
\arguments{
\item{data}{currently support \code{eg} and \code{vs}}

\item{reps}{the number of replicates planned. (should be tabled by TESTCODE, PERIOD and PHOUR). If \code{NULL}, \code{guess_reps} will be invoked to guess the number of replicates.}

\item{printed}{whether to print the observations having issues at the console.}
}
\value{
a data frame containing subject and issues at given period and protocol hours
}
\description{
check data qualities, reporting missing and RECHECKS
}
\examples{
d1 <- check_data(eg)
}
\seealso{
\code{guess_reps}.
}

