% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_quality.R
\name{check_eg}
\alias{check_eg}
\title{check for possile missing or recheck measurements.}
\usage{
check_eg(eg)
}
\arguments{
\item{eg}{the dataset eg.}
}
\value{
a data frame containing observations with possible issues.
}
\description{
Check the possible issues with ECG data
}

