% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_quality.R
\name{check_vs}
\alias{check_vs}
\title{check for possile missing or recheck measurements.}
\usage{
check_vs(vs)
}
\arguments{
\item{vs}{The dataset vs.}
}
\value{
a data frame containing observations with possible issues.
}
\description{
Check the possible issues with vital signs data
}

