% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab.R
\name{clean_lab}
\alias{clean_lab}
\title{Clean the lab test data set.}
\usage{
clean_lab(lab, ex, included)
}
\arguments{
\item{lab}{the dataset returned by \code{\link{create_lab}}}

\item{ex}{the ex data set.}

\item{included}{the included data set.}
}
\value{
lab data with necessary variables kept for further summarization.
}
\description{
clean the lab data, handling rechecks/unscheduled included.
}
\seealso{
\code{\link{create_lab}} and \code{\link{create_included}}
}

