% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{count_percent}
\alias{count_percent}
\title{add percentage to a frequency table}
\usage{
count_percent(data, var1, var2, digit_keep = 3)
}
\arguments{
\item{data}{the data should contain at least 2 columns}

\item{var1}{which column should be the group}

\item{var2}{which columns are used to calculate percentage (could be a vector)}

\item{digit_keep}{how many digits should be kept.}
}
\value{
a data frame
}
\description{
count to percent
}
\examples{
trait = rep(c("A", "B", "C"), 3); subtype = paste(trait, rep(1:3, each=3), sep = "")
data <- data.frame(trait, subtype, count1 = rpois(9, 5), count2 = rpois(9, 10)) 
count_percent(data, var1= 1, var2 = 3:4)

r1 <- get_summary_stats(eg, group = "PERIOD", var = "EG_TEST")
count_percent(r1, var1 = 1, var2 = 3:5)
}

