% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_ae.R
\name{create_aet}
\alias{create_aet}
\title{do what createaet does.}
\usage{
create_aet(ae, ex, included, improv = 99)
}
\arguments{
\item{ae}{the dataset ae read from sas}

\item{ex}{the dataset ex read from sas}

\item{included}{the dataset included from sas, can be created using \code{new_create_included()}}

\item{improv}{the same argument as the sas macro \code{createaet()}}
}
\value{
aet the data
}
\description{
createaet.
}

