% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{create_baseline}
\alias{create_baseline}
\title{create baseline indicator.}
\usage{
create_baseline(data, var_identifier = "_TEST")
}
\arguments{
\item{data}{currently support \code{vs}, \code{eg} and \code{lb_cq}.}

\item{var_identifier}{a string that can be used to identify the variable name e.g.(\code{VS_TEST, EG_TEST, LB_TEST})}
}
\value{
a data frame with an extra column \code{status} whose value could be one of \code{BASELINE}, \code{POSTDOSE} and \code{PREDOSE (NOT BASELINE)}
}
\description{
create an extra column inidcating whether the current row should be considered as baseline.
}
\seealso{
\code{\link{guess_base_phour}}
}

