% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_dem.R
\name{create_dem}
\alias{create_dem}
\title{create the demographics data.}
\usage{
create_dem(dm, ex, vs, included)
}
\arguments{
\item{dm}{the dm data set}

\item{ex}{the ex data set}

\item{vs}{the vs data set}

\item{inlcuded}{the included data set created by \code{create_included}}
}
\value{
a data frame
}
\seealso{
\code{\link{create_included}}
}

