% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_included.R
\name{create_included}
\alias{create_included}
\title{do what new_create_included does.}
\usage{
create_included(ex, dm, cr, ds)
}
\arguments{
\item{ex}{the dataset ex read from sas}

\item{dm}{the dataset dm read from sas}

\item{cr}{the dataset cr read from sas}

\item{ds}{the dataset ds read from sas}
}
\value{
the included data set
}
\description{
create_included.
}

