% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab.R
\name{create_lab}
\alias{create_lab}
\title{create the lab test data set.}
\usage{
create_lab(lb_cq, included)
}
\arguments{
\item{lb_cq}{the dataset lb_cq read from sas}

\item{included}{the included data set created by \code{\link{create_included}}}
}
\value{
lab data with necessary variables kept for further summarization.
}
\description{
create the lab data
}

