% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab.R
\name{create_labshift}
\alias{create_labshift}
\title{prepare data for lab shift table.}
\usage{
create_labshift(lb_cq, included, ex)
}
\arguments{
\item{lb_cq}{the dataset lb_cq read from sas}

\item{included}{the dataset created by \code{create_included}. This is the same as the data "included" from sas}

\item{ex}{the dataset ex read from sas}
}
\value{
a data frame
}
\description{
Clean the lb_cq data (for lab shif purpose)
}
\seealso{
\code{\link{create_included}}
}

