% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab_shift.R
\name{create_lb_cq}
\alias{create_lb_cq}
\title{prepare data for lab shift table.}
\usage{
create_lb_cq(lb_cq, included, ex)
}
\arguments{
\item{lb_cq}{the dataset lb_cq read from sas}

\item{included}{the dataset created by \code{create_included()}. This is the same as the data "included" from sas}

\item{ex}{the dataset ex read from sas}
}
\value{
a data frame
}
\description{
Clean the lb_cq data
}
\seealso{
\code{\link{create_included}}
}

