% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_included.R
\name{create_seq}
\alias{create_seq}
\title{Create SEQ}
\usage{
create_seq(data, included)
}
\arguments{
\item{data}{either \code{eg}, \code{vs} or \code{lb_cq}}

\item{included}{the included data set created by \code{\link{create_included}}}
}
\value{
the same data set with one more column if SEQ is created, otherwise returns the input data.
}
\description{
Create SEQ variable (treatment) if it's not already in the data set.
}

