% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{create_threshold}
\alias{create_threshold}
\title{create threshold table (for flagging variable purpose).}
\usage{
create_threshold(flagvar, lower = -Inf, upper = Inf, include_lower = T,
  include_upper = F, flg_label = "#", add2existing = F, thresh = NULL)
}
\arguments{
\item{flagvar}{the variable to be flagged}

\item{lower}{lower bound. Set to be \code{-Inf} if none.}

\item{upper}{upper bound. Set to be \code{Inf} if none.}

\item{include_lower}{whether the lower bound should be included, \code{TRUE} by default}

\item{include_upper}{whether the upper bound should be included, \code{FALSE} by default}

\item{flg_label}{what label will be used to flag the variable}

\item{add2existing}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, must specify thresh.}

\item{thresh}{a data frame with three columns: first column, the variable to be flagged, second column, lower bound (if any); third column, upper bound (if any).}
}
\value{
a data frame
}
\description{
create threshold data frame
}
\examples{
a <- create_threshold(flagvar = "QTCF", lower = 0, upper = 20, include_lower = FALSE, include_upper = TRUE,  flg_label = "^")
b <- create_threshold(flagvar = "PR",  upper = 430, flg_label = "*", add2existing = TRUE, thresh = a)
c <- create_threshold(flagvar = "QRs",  lower = 100, flg_label = "\%",add2existing = TRUE, thresh = b)
print(a); print(b);print(c);
}

