% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_vs.R
\name{create_vs}
\alias{create_vs}
\title{create the vital sign data set.}
\usage{
create_vs(vs, included)
}
\arguments{
\item{vs}{the dataset dm read from sas}

\item{included}{the included data set created by \code{\link{create_included}}}
}
\value{
vs data with necessary variables kept for further summarization.
}
\description{
create the vs data
}

