% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_dem.R
\name{dem_listing}
\alias{dem_listing}
\title{demographics listing.}
\usage{
dem_listing(dm, ex, vs, included)
}
\arguments{
\item{dm}{the dm data set}

\item{ex}{the ex data set}

\item{vs}{the vs data set}

\item{inlcuded}{the included data set created by \code{new_create_included}}
}
\value{
a data frame
}
\description{
Summarize the demographic data
}
\seealso{
\code{\link{new_create_included}}
}

