% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_ex.R
\name{ex_start}
\alias{ex_start}
\title{Get the start time of each treatment}
\usage{
ex_start(ex)
}
\arguments{
\item{ex}{the dataset ex read from sas}
}
\value{
a data frame specifying the start time of each treatment
}
\description{
Get the start time of each treatment
}

