% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{flg_var}
\alias{flg_var}
\title{label the variables of interest.}
\usage{
flg_var(data, thresh, prefix = "flg", oor = rep(F, nrow(thresh)))
}
\arguments{
\item{data}{the dataset to be flagged.}

\item{thresh}{a data frame returned by \code{create_threshold()}}

\item{prefix}{how will the flag variable be named (by adding a prefix).}

\item{oor}{is the current variable to be labeled for out of range?}
}
\value{
the data with variables flagged
}
\description{
falg variables by threshold
}
\examples{
a <- create_threshold(flagvar = "PR", lower = 150, upper = 200, include_lower = FALSE, include_upper = TRUE,  flg_label = "^")
b <- create_threshold(flagvar = "QTCF",  upper = 430, flg_label = "*", add2existing = TRUE, thresh = a)
data <- data.frame(PR = rnorm(10, 175, 25), QTCF = rnorm(10, 440, 20))
flg_var(data, thresh=b, oor=c(TRUE, FALSE))

}

