% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{format_time}
\alias{format_time}
\title{Format the time variable in a data set}
\usage{
format_time(data, date = "AE_STDT", time = "AE_STTM")
}
\arguments{
\item{data}{the data which contains the time variables}

\item{date}{the date variable}

\item{time}{the time (in seconds)}
}
\value{
the same data set but \code{time} has been changed to "hh:mm:ss" format.
}
\description{
Format the time variable in a data set
}

