% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_dem.R
\name{guess_race_age}
\alias{guess_race_age}
\title{race and ethnicity indicator, age.}
\usage{
guess_race_age(dm, included)
}
\arguments{
\item{dm}{the dataset dm read from sas}

\item{ex}{the dataset ex read from sas}
}
\value{
a data frame with additional columns listed as follows
\item{race}{the race of the subject}
\item{ethnic}{has two levels, "NOT HISPANIC OR LATINO" and "HISPANIC OR LATINO".}
\item{EX_TRT_C}{the treatment groups}
\item{ptno}{convert CLIENTID to numerical values of subject number}
\item{age}{Age calculated from start date of treatment}
}
\description{
create necessary variable for the demographic summary tables and listing.
}

