% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{guess_reps}
\alias{guess_reps}
\title{guess the number of replicates}
\usage{
guess_reps(data, var_identifier = "_TEST")
}
\arguments{
\item{data}{data of \code{eg} (created by \code{\link{create_eg}}) or \code{vs} (created by \code{\link{create_vs}})}

\item{var_identifier, }{which columns is the variable for test code names.}
}
\value{
a data frame containing number of reps per PERIOD/PHOUR/TESTCODE
}
\description{
guess the number of replicates for each SEQ/PERIOD/PHOUR/TESTCODE
}
\examples{
eg0 <- create(eg, included)
d1 <- guess_reps(eg0)
}

