% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{guess_test}
\alias{guess_test}
\title{guess the name of the test}
\usage{
guess_test(data, var_identifier = "_TEST")
}
\arguments{
\item{data}{currently support \code{eg} and \code{vs}}

\item{var_identifier, }{which columns is the variable for test code names.}
}
\value{
the test code
}
\description{
guess the name of the test code: it should end with var_identifier.
}
\examples{
guess_test(eg)
}

