% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{keep_non_empty}
\alias{keep_non_empty}
\title{remove empty rows and empty columns of a data set.}
\usage{
keep_non_empty(data, pattern = NA, return_truncated_data = F)
}
\arguments{
\item{data}{the dataset}

\item{pattern}{the pattern to be considered as empty. \code{NA} by default.}

\item{return_truncated_data}{Controls the output.}
}
\value{
a data frame with empty rows and empty columns removed,  or a list containing the following elements.
\item{keep_rows}{a vector of \code{TRUE} (for keeping row) or \code{FALSE} (otherwise)}
\item{keep_rows}{a vector of \code{TRUE} (for keeping column) or \code{FALSE} (otherwise)}
}
\description{
remove empty rows and empty columns of a data set
}
\examples{
data1 <- data.frame(x = rep(c("A", NA, "C"), each = 3), y = NA, z = c(1:3, NA, 5:9) )
keep_non_empty(data1, return_truncated_data = T)   
d1 <- keep_non_empty(data1)
}

