% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab.R
\name{lab_oor}
\alias{lab_oor}
\title{Find OOR}
\usage{
lab_oor(lab)
}
\arguments{
\item{lab}{the data returned by \code{\link{create_lab}}.}
}
\value{
a list containg OOR.
}
\description{
Find the out-of-range values
}

