% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_ae.R
\name{listing_ae}
\alias{listing_ae}
\title{list Adverse Envent}
\usage{
listing_ae(aet, type = 1)
}
\arguments{
\item{aet}{created by \code{create_aet}}

\item{type}{an indicator. Should I list \code{ae1 (type = 1)}, \code{ae2 (type = 2)} or \code{ae3 (type = 3)}?}
}
\value{
a data frame
}
\description{
list the ae
}
\seealso{
\code{\link{create_aet}}
}

