% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_con.R
\name{listing_concom}
\alias{listing_concom}
\title{list concomitant.}
\usage{
listing_concom(cm)
}
\arguments{
\item{cm}{the dataset cm read from sas}
}
\value{
the data with related columns
}
\description{
list concomitant medication data
}

