% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_dem.R
\name{listing_dem}
\alias{listing_dem}
\title{demographics listing.}
\usage{
listing_dem(dmt)
}
\arguments{
\item{dmt}{the data set created by \code{create_dem}.}
}
\value{
a data frame
}
\description{
Summarize the demographic data
}
\seealso{
\code{\link{new_create_included}} and \code{\link{create_dem}}
}

