% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{negative_hour}
\alias{negative_hour}
\title{find which variables have negative hours, note if "negativeHour == F" that means we should use screen or other (PHOUR) as baseline}
\usage{
negative_hour(temp1)
}
\arguments{
\item{temp1}{currently support \code{vs}, \code{eg} and \code{lb_cq}.}
}
\value{
a data frame where by-subject baseline hour is determined.
}
\description{
find the baseline hour for each test category
}
\examples{
hours <- negative_hour(eg)
}

