% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab.R
\name{normal_range}
\alias{normal_range}
\title{Find normal range for lab test}
\usage{
normal_range(lab)
}
\arguments{
\item{lab}{the dataset generated by \code{create_lab}}
}
\value{
a data frame containing test codes and their corresponding normal range
}
\description{
Find the normal range of each test code
}

