% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_eg.R
\name{replicate_average}
\alias{replicate_average}
\title{get the averanges of the replicates}
\usage{
replicate_average(data, var = c("HR", "PR", "QRS", "QT", "QTCF"),
  prefix = "Base")
}
\arguments{
\item{data}{an object returned from \code{replicate_clean}}

\item{var}{the variables used to do the calculation}
}
\value{
the averages
}
\description{
Get the averages of the replicates
}
\seealso{
\code{\link{replicate_clean}}
}

