% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{replicate_check}
\alias{replicate_check}
\title{check the replicates and issue a message if there's issue}
\usage{
replicate_check(data, reps = NULL, printed = TRUE)
}
\arguments{
\item{data}{data of \code{eg} (created by \code{\link{create_eg}}) or \code{vs} (created by \code{\link{create_vs}})}

\item{reps}{the number of replicates planned. (should be tabled by TESTCODE, SEQ, PERIOD and PHOUR). If \code{NULL}, \code{guess_reps} will be invoked to guess the number of replicates.}

\item{printed}{whether to print the observations having issues at the console.}
}
\value{
a data frame containing subject and issues at given period and protocol hours
}
\description{
check data qualities, reporting missing and RECHECKS
}
\examples{
eg0 <- create_eg(eg, incldued)
d1 <- replicate_check(eg1)
}
\seealso{
\code{guess_reps}.
}

