% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_eg.R
\name{replicate_clean}
\alias{replicate_clean}
\title{get the cleaned replicates by combining the "clean" data and "dirty" data}
\usage{
replicate_clean(data, rm_row = NULL)
}
\arguments{
\item{data}{an object returned from \code{replicate_eg}}

\item{rm_row}{a vector of integers specifying which rows should be removed from the dirty data.}
}
\value{
the cleaned replicates
}
\description{
get cleaned replicates
}
\seealso{
\code{\link{replicate_eg}}
}

