% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{replicate_data}
\alias{replicate_data}
\title{get the replicates for eg or vs data.}
\usage{
replicate_data(data, reps = NULL)
}
\arguments{
\item{data}{the data set created by \code{\link{create_eg}} (for ecg data) or by  \code{\link{create_vs}} (for vital signs data).}

\item{reps}{specifies the structures of reps. If not specified, \code{guess_resp} will be invoked. Default set to be \code{NULL}.}
}
\value{
a list 
 \item{data_clean}{the subjects containing correct number of resplicates}
 \item{data_dirty}{the subjects that have different number of replicates than desired}
}
\description{
get the replicates for each protocol hour (PHOUR)
}
\seealso{
\code{\link{create_eg}}, \code{\link{guess_resp}} and  \code{\link{replicate_check}}
}

