% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_eg.R
\name{replicate_eg}
\alias{replicate_eg}
\title{get the replicates for eg data.}
\usage{
replicate_eg(eg, reps = 3)
}
\arguments{
\item{eg}{the data set eg.}

\item{reps}{how many replicates will be measured for each protocol hour.}

\item{isBase}{Is this for finding baseline? Default set to be \code{TRUE}.}
}
\value{
a list 
 \item{data_clean}{the subjects containing correct number of resplicates}
 \item{data_dirty}{the subjects that have different number of replicates than desired}
}
\description{
get the replicates for each protocol hour (PHOUR)
}
\seealso{
\code{\link{create_eg}}
}

