% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{round_df}
\alias{round_df}
\title{round data frame}
\usage{
round_df(df, digits = 3)
}
\arguments{
\item{df}{the data frame}

\item{digits}{how many digits you want to keep}
}
\value{
a data frame with numerical columns rounded.
}
\description{
round a data frame if the column is numerical
}

