% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_ae.R
\name{summary_ae_sev}
\alias{summary_ae_sev}
\title{Adverse event by severity}
\usage{
summary_ae_sev(data, group = "STUDYNO", var = "EX_TRT_C", by = "subject")
}
\arguments{
\item{data}{the data set created by \code{\link{create_aet}}}

\item{group}{choose a varialbe that has all values identical (e.g., "STUDYNO").}

\item{var}{the treatment variable name, "EX_TRT_C" for example}

\item{by}{has two options, either "\code{event}" or "\code{subject}". If \code{by = "subject"}, the subject will be counted under the highest severity level.}
}
\value{
a list
\item{pt}{frequency table by preferred term}
\item{total}{frequency table by treatment}
}
\description{
summarize adverse events by severity
}

