% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_dem.R
\name{summary_dem}
\alias{summary_dem}
\title{demographic summary.}
\usage{
summary_dem(dmt, group = "EX_TRT_C", na.rm = TRUE)
}
\arguments{
\item{dmt}{the data set created by \code{create_dem}.}

\item{group}{which variable name would be used to calculate summary statistics?}

\item{na.rm}{should missing values be included?  \code{TRUE} by default.}
}
\value{
a data frame
}
\description{
Summarize the demographic data
}
\examples{
included <- new_create_included(ex, dm, cr, ds)
dmt <- create_dem(dm, ex, vs, included)
summary_dem(dmt, group = "SEQ")   # the summary by group
summary_dem(dmt, group = "SPONSOR")  # to get the overall summary
}
\seealso{
\code{\link{new_create_included}} and \code{\link{create_dem}}
}

