% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab.R
\name{summary_lab}
\alias{summary_lab}
\title{summary statistics for lab shift.}
\usage{
summary_lab(lb_clean, digits = 3)
}
\arguments{
\item{lb_clean}{the dataset returned by \code{clean_lab}.}

\item{digits}{how many digits should be kept for the final data.}
}
\value{
the summary statistics by test code, time point and by treatment.
}
\description{
Summarize the lab statistics
}
\examples{
included <- create_included(ex, dm, cr, ds)
lab <- create_lab(lb_cq)
lb_clean <- clean_lab(lab, included= included, ex= ex)
s1 <- summary_lab(clean_lab)  #not urinalysis

}
\seealso{
\code{\link{clean_lab}}
}

