% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab.R
\name{summary_labshift}
\alias{summary_labshift}
\title{create the lab shift table.}
\usage{
summary_labshift(lb_clean)
}
\arguments{
\item{lb_clean}{the dataset returned by \code{clean_lab}.}
}
\value{
the shift table
}
\description{
Create the lab shift table.
}
\examples{
included <- create_included(ex, dm, cr, ds)
lab <- create_lab(lb_cq)
lb_clean <- clean_lab(lab, included, ex)
lbshift <- summary_labshift(lb_clean)  

}
\seealso{
\code{\link{clean_lab}}
}

