% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_ae.R
\name{summary_teae}
\alias{summary_teae}
\title{teae by number of subjects or by number of events}
\usage{
summary_teae(data, group = "STUDYNO", var = "EX_TRT_C", by = "subject")
}
\arguments{
\item{data}{the data set created by \code{\link{create_aet}}}

\item{group}{choose a varialbe that has all values identical (e.g., "STUDYNO").}

\item{var}{the treatment variable name, "EX_TRT_C" for example}

\item{by}{has two options, either "\code{event}" or "\code{subject}"}
}
\value{
a list
\item{ae}{frequency table for teae}
\item{soc}{summarization by system organ class and by preferred term}
}
\description{
summarize treatment-emergent adverse event
}

