% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_dem.R
\name{weight_height_bmi}
\alias{weight_height_bmi}
\title{read BMI Weight and Height info from Screening stage.}
\usage{
weight_height_bmi(vs)
}
\arguments{
\item{vs}{the vs sas data}
}
\value{
a data frame containing BMI, WEIGHT and HEIGHT from admission stage
}
\description{
extract BMI, weight and height from vs.sas7bat.
}

