% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_eg.R
\name{change_from_base}
\alias{change_from_base}
\title{calculate the change from baselines}
\usage{
change_from_base(data, var = c("HR", "PR", "QRS", "QT", "QTCF"))
}
\arguments{
\item{var}{the variables used to do the calculation}

\item{data.}{An object returned from \code{replicate_clean}}
}
\value{
a data frame
}
\description{
Get the averages of the replicates
}
\seealso{
\code{\link{replicate_clean}}

\code{\link{replicate_eg}}

\code{\link{create_eg}}
}

