% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_eg.R
\name{create_eg}
\alias{create_eg}
\title{create variable columns for the ECGs parameters.}
\usage{
create_eg(eg, included)
}
\arguments{
\item{eg}{the dataset eg read from sas}

\item{included}{the included data set created by \code{\link{create_included}}}
}
\value{
eg1 data with necessary variables kept for further summarization.
}
\description{
dcast the ecg variables
}

