% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_vs.R
\name{create_vs}
\alias{create_vs}
\title{create the vital sign data set.}
\usage{
create_vs(vs)
}
\arguments{
\item{vs}{the dataset dm read from sas}
}
\value{
the transposed data
}
\description{
dcast the vs variables
}

