% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_dem.R
\name{dem_summary}
\alias{dem_summary}
\title{demographic summary.}
\usage{
dem_summary(dm, ex, vs)
}
\arguments{
\item{dm}{the dm data set}

\item{ex}{the ex data set}

\item{vs}{the vs data set}
}
\value{
a data frame
}
\description{
Summarize the demographic data
}

