% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{find_base_phour}
\alias{find_base_phour}
\title{find the baseline hour for each test}
\usage{
find_base_phour(data, var_identifier = "_TEST")
}
\arguments{
\item{data}{currently support \code{vs}, \code{eg} and \code{lb_cq}.}

\item{var_identifier}{a string that can be used to identify the variable name e.g.(\code{VS_TEST, EG_TEST, LB_TEST})}
}
\value{
a data frame where by-subject baseline hour is determined.
}
\description{
find the baseline hour for each test category
}
\examples{
# the following two will give you exactly the same result
r1 <- find_base_phour(vs)
r2 <- find_base_phour(vs, var_identifier = "VS_TEST")

# If you want to find baseline hours for eg, the following two methods are equivalent
r3 <- find_base_phour(eg)
r4 <- find_base_phour(eg, var_identifier = "EG_TEST")

}

