% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_quality.R
\name{guess_reps}
\alias{guess_reps}
\title{guess the number of replicates}
\usage{
guess_reps(data, var_identifier = "_TEST")
}
\arguments{
\item{data}{currently support \code{eg} and \code{vs}}

\item{var_identifier, }{which columns is the variable for test code names.}
}
\value{
a data frame containing number of reps per PERIOD/PHOUR/TESTCODE
}
\description{
guess the number of replicates for each PERIOD/PHOUR/TESTCODE
}
\examples{
d1 <- guess_reps(eg)
}

