% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab_oor.R
\name{lab_oor}
\alias{lab_oor}
\title{print OOR values for all LB_CAT}
\usage{
lab_oor(lb_cq, ex, var_per_block = 5, digit_keep = 2)
}
\arguments{
\item{lb_cq}{the dataset lb_cq read from sas}

\item{ex}{the dataset ex, used to get treatment information.}

\item{var_per_block}{how many variables to be shown in each block}

\item{digit_keep}{the digit kept in the output. see \code{\link[knitr]{kable()}} for more details}
}
\description{
print out the OOR values for all LB_CAT automatically. This function is built on top of customized_lab_rept()
}

