% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab_oor.R
\name{normal_range}
\alias{normal_range}
\title{Find normal range}
\usage{
normal_range(laborig)
}
\arguments{
\item{laborig}{the dataset generated by \code{create_laborig}}
}
\value{
a data frame containing test codes and their corresponding normal range
}
\description{
Find the normal range of each test code
}

