% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_eg.R
\name{replicate_average}
\alias{replicate_average}
\title{get the averanges of the replicates}
\usage{
replicate_average(data, var = c("HR", "PR", "QRS", "QT", "QTCF"),
  prefix = "Base")
}
\arguments{
\item{data}{an object returned from \code{replicate_clean}}

\item{var}{the variables used to do the calculation}

\item{prefix}{give a prefix to \code{var} so the names are changed after calculating the averages}
}
\value{
the averages
}
\description{
Get the averages of the replicates
}
\examples{
eg2 <- replicate_eg(eg, reps = 3) # find the triplicates
eg_prob <- eg2$data_dirty         # need manual check
# the following rows should be removed
rows_removed <- c(2, 4, 7, 13, 14, 15, 19, 25, 28, 32, 37, 40, 44, 52, 57, 59, 64, 65)
eg3 <- replicate_clean(eg2, rows_removed)
ave <- replicate_average(eg3, prefix = "Ave")
}
\seealso{
\code{\link{replicate_clean}}
}

