% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{replicate_clean}
\alias{replicate_clean}
\title{get the cleaned replicates by combining the "clean" data and "dirty" data}
\usage{
replicate_clean(data, rm_row = NULL)
}
\arguments{
\item{data}{an object returned from \code{replicate_data} (works for vs and eg)}

\item{rm_row}{a vector of integers specifying which rows should be removed from the dirty data.}
}
\value{
the cleaned replicates
}
\description{
get cleaned replicates
}
\examples{
eg2 <- replicate_data(eg) # step 1: find the triplicates
eg_prob <- eg2$data_dirty         # need manual check
# the following rows should be removed
rows_removed <- c(2, 4, 7, 13, 14, 15, 19, 25, 28, 32, 37, 40, 44, 52, 57, 59, 64, 65)
eg3 <- replicate_clean(eg2, rows_removed)
}
\seealso{
\code{\link{replicate_data}}
}

