% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab.R
\name{summary_lab}
\alias{summary_lab}
\title{create the lab shift table.}
\usage{
summary_lab(laborig, included, digits = 3)
}
\arguments{
\item{laborig}{the dataset returned by \code{create_laborig}.}

\item{included}{the dataset created by \code{create_included}. This is the same as the data "included" from sas}

\item{digits}{how many digits should be kept for the final data.}
}
\value{
the summary statistics by test code, time point and by treatment.
}
\description{
Summarize the lab statistics
}
\examples{
included <- create_included(ex, dm, cr, ds)
laborig <- create_labshift(lb_cq, included, ex)
s1 <- summary_lab(laborig, included)  #not urinalysis

}
\seealso{
\code{\link{create_laborig}}
}

