% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab_shift.R
\name{summary_labshift}
\alias{summary_labshift}
\title{create the lab shift table.}
\usage{
summary_labshift(lb_cq, included, ex, UA = FALSE)
}
\arguments{
\item{lb_cq}{the dataset lb_cq read from sas}

\item{included}{the dataset included from sas (need variable SEQ)}

\item{ex}{the dataset ex read from sas}

\item{UA}{whether to produce shift table for UA (should be done separately)}
}
\value{
the shift table
}
\description{
Create the lab shift table.
}
\examples{
included <- create_included(ex, dm, cr, ds)
others <- labshift(lb_cq, included, ex,  UA=FALSE)  #not urinalysis
UA <- summary_labshift(lb_cq, included, ex, UA=TRUE)       # UA

}

