% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdash_lab.R
\name{summary_labshift}
\alias{summary_labshift}
\title{create the lab shift table.}
\usage{
summary_labshift(laborig, UA = FALSE)
}
\arguments{
\item{laborig}{the dataset returned by \code{create_labshift}.}

\item{UA}{whether to produce shift table for UA (should be done separately)}
}
\value{
the shift table
}
\description{
Create the lab shift table.
}
\examples{
included <- create_included(ex, dm, cr, ds)
laborig <- create_labshift(lb_cq, included, ex)
others <- summary_labshift(laborig,  UA=FALSE)  #not urinalysis
UA <- summary_labshift(laborig, UA=TRUE)       # UA

}
\seealso{
\code{\link{create_labshift}}
}

